# oas3_allyoucanorder.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: "AllYouCanOrder"
  description: API for managing restaurant orders.
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Localhost
paths:
...
  /tavoliCliente/ordine:
    get:
      description: >-
          Gets the list of dishes on the array "ordine" of a specific "tavolo". 
      summary: View all dishes in the array ordine
      responses:
        '200':
          description: 'Collection of dishes'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Piatto'
              example:
                - nome: 'Gamberi fritti'
                  prezzo: 2.49
                  descrizione: '5 pezzi. Ingredienti: gamberi'
                  foto: 'www.gamberifritti.images.com'
                  stato: 'in consegna'
                - nome: 'California Roll'
                  prezzo: 2.79
                  descrizione: '5 pezzi. Ingredienti: riso, salmone crudo, philadelphia'
                  foto: 'www.gamberifritti.images.com'
                  stato: 'in peparazione' 
        '404':
            description: 'table not found'
    post:
      description: >-
          Questa post serve per inserire tutti gli elementi in tavolo.carrello nell'array tavolo.ordine e viene chiamata in carrello.js nel metodo "svuotaCarrello"
      summary: Sposta piatti da carrello a ordine
      requestBody:
        content:
            application/json:
                schema:
                  type: object
                  required:
                    - nome
                    - prezzo
                  properties:
                    nome:
                      type: string
                      description: 'Nome del piatto'
                    prezzo:
                      type: number
                      description: 'Prezzo del piatto'
                    descrizione:
                      type: string
                      description: 'Descrizione del piatto'
                    foto:
                      type: string
                      description: 'Indirizzo alla foto del piatto'
                example:
                    nome: 'Nigiri al salmone'
                    prezzo: 0
                    descrizione: 'Ingredienti: salmone e riso'
                    foto: 'www.foto.com'
      responses:
        '201':
          description: 'Piatto aggiunto all ordine'
        '404':
            description: 'Ristorante non trovato'
        '404':
            description: 'Tavolo non trovato'
  /tavoliCliente/ilMioTavolo:
    get:
      description: >-
          Gets the table which id is the one selected in the test section. This api is used for example for the state of the button "chiama cameriere"
      summary: View the selected table 
      responses:
        '200':
          description: 'get the selected table'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tavolo'
              example:
                - nome: 'tavolo 1'
                  ordine: 
                  - nome: 'Nigiri di Tonno'  
                    prezzo: 3.99
                    descrizione: '8 pezzi. Ingredienti: riso, tonno crudo'
                    foto: 'www.fotoDelPiatto.com'
                    stato: 'in preparazione'
                  - nome: 'Nigiri di salmone'  
                    prezzo: 3.99
                    descrizione: '8 pezzi. Ingredienti: riso, salmone crudo'
                    foto: 'www.fotoDelPiatto2.com'
                    stato: 'in consegna'
                  carrello: 
                  - nome: 'Temaki salmone'  
                    prezzo: 3.99
                    descrizione: '1 pezzo. Ingredienti: riso, salmone crudo'
                    foto: 'www.foto.com'
                  chiamato: false
        '404':
            description: 'table not found'
  /tavoliCliente/chiamaCameriere:          
    post:
      description: >-
        Sets tavolo.chiamato = true 
      summary: set tavolo.chiamato = true
      responses:
        '200':
          description: 'Table chiamato is modified. Link in the Location header'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tavolo'
              example:
                nome: 'Tavolo1'  
                chiamato: true
        '404':
            description: 'tavolo not found' 
        '404':
            description: 'ristorante not found'
  /tavoliCliente/mostraCarrello:
    get:
      description: >-
          Questa get serve per mostrare tutti gli elementi nell'array tavolo.carrello e viene chiamata in carrello.js nel metodo "mostraCarrello()"
      summary: 'Mostra i piatti nel carrello'
      responses:
        '201':
          description: 'Lista dei piatti nel carrello'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tavolo'
        '404':
            description: 'Tavolo non trovato'
            content:
              application/json:
                schema:
                 $ref: '#/components/schemas/Tavolo'
    post:
      description: >-
        Inserts the selected dish in the menu section to the cart
      summary: Add a new dish in the cart
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - prezzo
              properties:
                nome:
                  type: string
                  description: 'Name of the dish'
                prezzo: 
                  type: number
                  description: 'Price of the dish'
                descrizione: 
                  type: string
                  description: 'Description of the dish'
                foto: 
                  type: string
                  description: 'Path to the photo of the dish'
            example:
              nome: 'Sashimi salmone'  
              prezzo: 0.00
              descrizione: '5 pz'
              foto: 'https://example.com/photo.jpg'
      responses:
        '200':
          description: 'Dish added to the cart. Link in the Location header'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Piatto'
              example:
                nome: 'Nigiri di Tonno'  
                prezzo: 3.99
                descrizione: '8 pezzi. Ingredienti: riso, tonno crudo'
                foto: 'www.fotoDelPiatto.com'
        '404':
            description: 'table not found'  
...
  /tavoliCliente/SvuotaCarrello:
    delete:
      description: >-
          Questa delete serve per eliminare tutti gli elementi nell'array tavolo.carrello e viene chiamata in carrello.js nel metodo "svuotaCarrello"
      summary: Svuota il carrello
      responses:
        '201':
          description: 'Tutti i piatti rimossi dal carrello'
        '404':
            description: 'Ristorante non trovato'
          









#GESTIONE AUTENTICAZIONE
...
components:
  schemas:
    Piatto:
      type: object
      required: 
      - nome
      - prezzo
      properties: 
        nome: 
          type: string
          description: 'Name of the dish'
        prezzo: 
          type: number
          description: 'Price of the dish'
        descrizione: 
          type: string
          description: 'Description of the dish'
        foto: 
          type: string
          description: 'path to the photo of the dish'
        stato: 
          type: string
          description: 'status of the dish (in preparazione/consegnato/in consegna)'
      description: 'dishes in the database'
    Tavolo:
      type: array
      items: 
        type: object
        required:
        - nome
        properties:
          nome:
            type: string
            description: 'Table s name'
          ordine:
            type: array
            items: 
              $ref: '#/components/schemas/Piatto'
            description: 'Table s order'
          chiamato:
            type: string
            description: 'Table has call the waiter'
          carrello:
            type: array
            items: 
              $ref: '#/components/schemas/Piatto'
            description: 'piatti in the cart'
        description: 'Table s in database'
...